remove unused files, commented out code, console.logs, rename a couple files
