--- public/index-old.html
@@ -1,54 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-        <!-- Bootstrap CSS -->
-        <link rel="stylesheet" 
-            href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" 
-            integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" 
-            crossorigin="anonymous">
-        <!-- Google fonts -->
-        <link rel="preconnect" 
-            href="https://fonts.googleapis.com">
-        <link rel="preconnect" 
-            href="https://fonts.gstatic.com" crossorigin>
-        <link rel="stylesheet" 
-            href="https://fonts.googleapis.com/css2?family=Roboto+Condensed:wght@300&display=swap" >
-        <!-- CSS stylesheet -->
-        <link rel="stylesheet"
-            href="./assets/style/style.css">
-
-        <title>picklestack</title>
-    </head>
-    <body>
-        <nav class="py-4"><span><a href="#">Home</a> | <a href="#">Login</a> | <a href="#">About</a></span></nav>
-        <div class="jumbotron text-center bg-transparent">
-            <h1 class="display-3 my-3 text-lowercase">Picklestack</h1>
-            <p class="lead my-4">Pickleball court locator</p>
-            <hr class="my-3">
-            <p class="my-1">Enter location:</p>
-            <input class="text-center my-3" type="text" id="name" placeholder="City, state, zipcode..." autocomplete="off" spellcheck="off">
-            <p class="lead">
-                <a class="btn btn-primary btn-lg" href="#" role="button">Locate</a>
-            </p>
-        </div>
-
-        <footer><p>© picklestack 2023</p></footer>
-
-        <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
-            integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" 
-            crossorigin="anonymous">
-            </script>
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" 
-            integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" 
-            crossorigin="anonymous">
-            </script>
-        <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" 
-            integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" 
-            crossorigin="anonymous">
-            </script>
-        <script type="text/javascript" src="./assets/script/script.js"></script>
-    </body>
-</html>
