--- public/pages/login.html
@@ -1,45 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" 
-      href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" 
-      integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" 
-      crossorigin="anonymous">
-    <link rel="stylesheet"
-      href="../assets/style/style.css">
-
-    <title>Pickleball Locator - Login</title>
-  </head>
-  <body>
-    <header>
-      <h1>Pickleball Locator</h1>
-    </header>
-    <main>
-      <form class="login-container">
-        <label for="username">Username:</label>
-        <input type="text" id="username" class="username" placeholder="Your username..." spellcheck="off">
-        <label for="password">Password:</label>
-        <input type="password" id="password" class="password" placeholder="Your password..." spellcheck="off">
-        <button type="login" class="login">Login</button>
-      </form>
-    </main>
-    <!-- jQuery first, then Popper.js, then Bootstrap JS -->
-    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
-      integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" 
-      crossorigin="anonymous">
-      </script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" 
-      integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" 
-      crossorigin="anonymous">
-      </script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" 
-      integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" 
-      crossorigin="anonymous">
-      </script>
-    <script type="text/javascript" src="../assets/script/script.js"></script>
-  </body>
-</html>
