+++ views/location.handlebars
@@ -8,8 +8,6 @@
     <div class="card-text">
       <p class="my-4">{{location.description}}</p>
     </div>
-
-    {{!-- <a href="/{{location.id}}" class="stretched-link"></a> --}}
   </div>
 </div>
 {{/each}}
