+++ views/search.handlebars
@@ -24,4 +24,4 @@
 </div>
 {{/each}}
 
-<script type="text/javascript" src="/assets/js/script.js"></script>
+<script type="text/javascript" src="/assets/js/index.js"></script>
